// src/app/core/models/prensip-response.ts
import { KullaniciOzet } from './kullanici-ozet';

export interface PrensipResponse {
  id: number;
  version: number;
  maddeNo: number;
  maddeAdi: string;
  icerikSurum: number;
  icerik: string | null;
  durumu: string;
  yururlukteMi: boolean;
  ekleyenKullanici: KullaniciOzet | null;
  eklemeTarihi: string | null;
  degistireKullanici: KullaniciOzet | null;
  degistirmeTarihi: string | null;
  onaylayan: KullaniciOzet | null;
  onayTarihi: string | null;
}